image uploder server

+++ src/index.ts
@@ -2,12 +2,13 @@ import { config } from "dotenv";
 import express, { Application, NextFunction, Request, Response } from "express";
 import fileUpload from "express-fileupload";
 import path from "path";
-
+import morgan from "morgan";
 import fs from "fs";
 
 import rimraf from "rimraf";
 
 config();
+
 
 const app: Application = express();
 
@@ -15,6 +16,7 @@ app.get("/", (req: Request, res: Response, next: NextFunction) => {
   res.send("Dev Challenges Image Uploader");
 });
 
+app.use(morgan("dev"));
 app.use(fileUpload());
 
 // app.use("/tmp", express.static(path.join(__dirname, "../tmp")));
@@ -37,34 +39,33 @@ app.get("/tmp/:filename", (req: Request, res: Response, next: NextFunction) => {
   // display file
 
   res.sendFile(filepath);
-
 });
-
 
 // delete files after 1 hour
 const uploadsDir = path.join(__dirname, "../tmp");
 
-
 setInterval(() => {
-fs.readdir(uploadsDir, function (err, files) {
-  console.log(files);
-  files.forEach(function (file, index) {
-    fs.stat(path.join(uploadsDir, file), function (err, stat) {
-      let endTime, now;
-      if (err) {
-        return console.error(err);
-      }
-      now = new Date().getTime();
-      // 15 minutes
-      endTime = new Date(stat.ctime).getTime() + 15 * 60 * 1000;
-      if (now > endTime) {
-        return rimraf(path.join(uploadsDir, file)).then(() => {
-          console.log("Successfully deleted");
+  fs.readdir(uploadsDir, function (err, files) {
+    console.log(files);
+    if (files !== undefined) {
+      files.forEach(function (file, index) {
+        fs.stat(path.join(uploadsDir, file), function (err, stat) {
+          let endTime, now;
+          if (err) {
+            return console.error(err);
+          }
+          now = new Date().getTime();
+          // 15 minutes
+          endTime = new Date(stat.ctime).getTime() + 15 * 60 * 1000;
+          if (now > endTime) {
+            return rimraf(path.join(uploadsDir, file)).then(() => {
+              console.log("Successfully deleted");
+            });
+          }
         });
-      }
-    });
+      });
+    }
   });
-});
 }, 60000);
 
 app.post("/api/upload", (req: Request, res: Response, next: NextFunction) => {
@@ -72,12 +73,16 @@ app.post("/api/upload", (req: Request, res: Response, next: NextFunction) => {
     return res.status(400).json({ msg: "No file uploaded" });
   }
 
-  try{
+  // check wether the directory exists or not
+  if (!fs.existsSync(path.join(__dirname, "../tmp"))) {
+    fs.mkdirSync(path.join(__dirname, "../tmp"));
+  }
 
+  try {
     const file: any = req.files.file;
-    
+
     console.log(file);
-    
+
     // random file name
     const fileName = Math.random().toString(36).substring(7);
 
@@ -91,7 +96,7 @@ app.post("/api/upload", (req: Request, res: Response, next: NextFunction) => {
     const fileFinalName = `${fileFirstName}-${fileName}.${fileExtension}`;
 
     const filepath = path.join(__dirname, "../tmp", fileFinalName);
-    
+
     file.mv(filepath, (err: any) => {
       if (err) {
         console.error(err);
@@ -101,13 +106,12 @@ app.post("/api/upload", (req: Request, res: Response, next: NextFunction) => {
       // current domain name
       const domainName = req.protocol + "://" + req.get("host");
       console.log(domainName);
-      
+
       res.json({
-        fileName: fileFirstName,
-        filePath: `${domainName}/tmp/${fileFinalName}`,
+        url: `${domainName}/tmp/${fileFinalName}`,
       });
     });
-  }catch(err){
+  } catch (err) {
     console.log(err);
     return res.status(500).json({ msg: "Server Error" });
   }
